fixed bug where complete wouldnt work

+++ lib/screens/checkbox.dart
@@ -92,11 +92,14 @@ class _CheckBoxNoteState extends State<CheckBoxNote> {
               boop = true;
               val = false;
             }
-            var tert = all.notifs[widget.id]!;
-            String ter = tert.id2;
-            all_notes.notifs.remove(tert.id);
-            tert.delete();
-            NotificationService().deleteNotif(ter);
+            if (all.notifs[widget.id] != null) {
+              var tert = all.notifs[widget.id]!;
+              String ter = tert.id2;
+              all_notes.notifs.remove(tert.id);
+              tert.delete();
+              NotificationService().deleteNotif(ter);
+            }
+
             all_notes.uncompleted.remove(item);
             final id = Localstore.instance.collection("notes").doc().id;
             all_notes.searchResults.remove(item);
@@ -116,6 +119,7 @@ class _CheckBoxNoteState extends State<CheckBoxNote> {
             item1.save();
 
             all_notes.searchResults.add(item1);
+
             Notifs notif1 = Notifs(
               id: id,
               id2: count.channelCounter.toString(),
@@ -162,6 +166,7 @@ class _CheckBoxNoteState2 extends State<CheckBoxNote2> {
         value: val,
         onChanged: (val) {
           setState(() {
+            print(all.items);
             store.Notes item = all.items[widget.id]!;
 
             if (colPick == const Color.fromARGB(255, 255, 254, 254)) {
